changes: validationFunctions file moved to util
